requirements.txt
dvc>=2.8.3,<3
joblib>=1.0.1,<2
jupyter>=1.0.0,<2
jupyter_contrib_nbextensions>=0.5.1,<1
matplotlib>=3.4.3,<4
numpy>=1.21.2,<2
pandas>=1.3.2,<2
pytest>=6.2.4,<7
python-box>=5.4.1,<6
pyyaml!=6.0.0,!=5.4.0,!=5.4.1 # pyyaml is broken with cython 3
scikit-learn>=0.24.2,<2
scipy>=1.7.1,<2
tqdm>=4.62.2,<5
python-dotenv~=1.0.1